refactor(web): break the api/ <-> data-layer import cycle (P3a)
Remove the runtime cycle between api/mutations/* and hooks/github/*: drop
the 3 barrel re-exports from hooks/github/mutations.ts (callers now import
the orchestrators from api/mutations/* directly), source STUDENT_CSV_FIELDS
from the pure util/rosterCsv (deleting the TDZ workaround), and relocate the
config-repo read primitives from api/github/queries.ts down into
github-core (hooks/github/configRepoReads.ts) so the dependency points one
way. Add an eslint-plugin-import-x import-x/no-cycle guard scoped to the
data layer so the cycle can't silently return.

First phase (P3a) of the web-architecture-boundaries plan; behavior-
preserving, no contract values changed.
