+++ web/src/api/mutations/assignments.ts
@@ -12,7 +12,7 @@ import {
   getClassroomJson,
   getCommit,
   getConfigRepoBranch,
-} from "../github/queries"
+} from "@/hooks/github/configRepoReads"
 import { getUser } from "@/hooks/github/queries"
 import { CONFIG_REPO, DEFAULT_BRANCH } from "@/util/configRepo"
 import { classroomTeamSlug } from "@/util/teamSlug"
