+++ web/src/pages/CreateAssignmentPage.tsx
@@ -11,7 +11,7 @@ import { EmptyRosterNotice } from "@/components/EmptyRosterNotice"
 import CreateAssignmentForm from "@/pages/assignments/CreateAssignmentForm"
 import { useDocumentTitle } from "@/hooks/useDocumentTitle"
 import { GitHubAPIError } from "@/hooks/github/errors"
-import { createAssignment } from "@/hooks/github/mutations"
+import { createAssignment } from "@/api/mutations/assignments"
 import { useGitHubClient } from "@/context/github/GitHubProvider"
 import { useToast } from "@/context/notifications/NotificationProvider"
 import { useActionActivityRegistry } from "@/context/actions/ActionActivityProvider"
